fix: 차트 대기VC 삭제 및 빈 차트일때 UI 수정
